Fixes for maxrects
Now it is able to run but code is spaghettified.
It for sure does not work, all bins are placed on the bottom right and
many of them are missing, it seems that they are out of the bin?

Additionally overlapped rects are removed and we probably remove items
with them!